Lint test targets with Whitaker
Add `--all-targets` to the local and CI Whitaker invocations so
`#[cfg(test)]` modules are checked by the same lint suite as Clippy.

Resolve the violations exposed by the wider target set by splitting large
modules, adding missing module documentation, and replacing panic-prone test
helper code with fallible control flow where appropriate.
