+++ src/connection_flags.rs
@@ -42,6 +42,7 @@ impl ConnectionFlags {
 
 #[cfg(test)]
 mod tests {
+    //! Tests for this module.
     use rstest::rstest;
 
     use super::*;
