+++ src/db/file_path.rs
@@ -134,6 +134,7 @@ where
 
 #[cfg(test)]
 mod tests {
+    //! Tests for this module.
     use rstest::rstest;
 
     use super::prepare_path;
