+++ src/protocol.rs
@@ -113,6 +113,7 @@ where
 
 #[cfg(test)]
 mod tests {
+    //! Tests for this module.
     use super::*;
 
     #[test]
