+++ src/transaction/writer.rs
@@ -192,6 +192,7 @@ where
 
 #[cfg(test)]
 mod tests {
+    //! Tests for this module.
     use std::io::Cursor;
 
     use tokio::io::{BufReader, BufWriter, duplex};
