+++ tests/wireframe_routing_bdd.rs
@@ -90,8 +90,9 @@ impl RoutingWorld {
 
 #[fixture]
 fn world() -> RoutingWorld {
-    ensure_server_binary_env(env!("CARGO_BIN_EXE_mxd-wireframe-server"))
-        .unwrap_or_else(|error| panic!("failed to configure wireframe test binary path: {error}"));
+    if let Err(error) = ensure_server_binary_env(env!("CARGO_BIN_EXE_mxd-wireframe-server")) {
+        panic!("failed to configure wireframe test binary path: {error}");
+    }
     RoutingWorld::new()
 }
