+++ tests/wireframe_xor_compat.rs
@@ -64,8 +64,9 @@ impl XorWorld {
 
 #[fixture]
 fn world() -> XorWorld {
-    ensure_server_binary_env(env!("CARGO_BIN_EXE_mxd-wireframe-server"))
-        .unwrap_or_else(|error| panic!("failed to configure wireframe test binary path: {error}"));
+    if let Err(error) = ensure_server_binary_env(env!("CARGO_BIN_EXE_mxd-wireframe-server")) {
+        panic!("failed to configure wireframe test binary path: {error}");
+    }
     let world = XorWorld::new();
     assert!(!world.is_skipped(), "world starts active");
     world
